+++ src/components/navbar/mobile/parts/buttons/ButtonsManagerM.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import ProfilePicAndNameNavM from '@src/components/navbar/mobile/parts/profileDataM/ProfilePicAndNameNavM';
-import ButtonsM from './ButtonsM';
+import LoggedButtonsM from './LoggedButtonsM';
+import AnonymusButtonsM from './AnonymousButtonsM';
 
 const ButtonsManagerM = ({ isLogged }: { isLogged: boolean }) => {
   return (
     <div className='flex flex-col ml-6 gap-4 mt-12'>
-      {isLogged && <ProfilePicAndNameNavM />}
-      <ButtonsM isLogged={isLogged} />
+      {isLogged ? <LoggedButtonsM /> : <AnonymusButtonsM />}
     </div>
   );
 };
